Separete the Recipe class from file entity
